Move AI types back into the AI folder
