+++ app/src/lib/ai/types.ts
