+++ app/src/lib/components/AISettings.svelte
@@ -6,7 +6,7 @@
 	import TextBox from './TextBox.svelte';
 	import WelcomeSigninAction from './WelcomeSigninAction.svelte';
 	import { AIService, GitAIConfigKey, KeyOption } from '$lib/ai/service';
-	import { OpenAIModelName, AnthropicModelName, ModelKind } from '$lib/aiTypes';
+	import { OpenAIModelName, AnthropicModelName, ModelKind } from '$lib/ai/types';
 	import { GitConfigService } from '$lib/backend/gitConfigService';
 	import RadioButton from '$lib/components/RadioButton.svelte';
 	import SectionCard from '$lib/components/SectionCard.svelte';
